Hello @peterbaile,

Thank you for sharing your work! I had a lot of fun reading your paper and appreciate the novel ideas you’ve brought to the table ( no pun intended :) )

I noticed that one of the limitations you mention is the scalability bottleneck of the MIP-based re-ranking approach: as the number of candidate tables grows, the MIP solve becomes infeasible. To address this, I’ve implemented a **Greedy + Local-Swap** heuristic that achieves nearly **10× faster** runtimes while maintaining comparable end-to-end performance. I thought maybe it could serve towards writing a v2 of the paper ;)

The script is `greedy.py`, I've also included a quick guide on how to execute it, so it should already work with the existing codebase. 

---

## Greedy + Local-Swap Heuristic

This method replaces the global MIP solve with a two-phase procedure:

1. **Greedy Selection**: incrementally build the table set by picking the candidate with the highest marginal gain.  
2. **Local Swap Refinement (optional, sometimes improves results, sometimes not)**: perform pairwise exchanges to recover from greedy mistakes.

Both phases optimize the same weighted objective over three signals:

$\lambda_c \sum_{p\in Q}\max_{i\in S}F_{p,i} + \lambda_j \sum_{i&lt;j,\,i,j\in S}J_{ij} + \lambda_r \sum_{i\in S}R_i$

Where:

- **Fine-grained coverage**: $F_{p,i} \in [0,1]$ is the maximum column-similarity between sub-question $p$ and table $i$  
- **Join compatibility**: $J_{ij} = \max_{k,\ell} C^{(i,j)}_{k,\ell}$ measures how well tables $i$ and $j$ can join  
- **Coarse relevance**: $R_i$ is the retrieval score (e.g., BM25 or bi-encoder) of table $i$

---

### Pseudo-Algorithm

```text
Input:
  T = {1,…,M}        ← candidate tables  
  Q = {1,…,P}        ← decomposed sub-questions  
  K                  ← max tables to select  
  λ_c, λ_j, λ_r      ← weights  
  τ                  ← max swap passes

Phase 1: Greedy Selection
-------------------------
1. For each i ∈ T:
     G_i ← λ_c * Σ_{p∈Q} F_{p,i}  +  λ_r * R_i

2. i* ← argmax_i G_i

3. S ← {i*}
   M_p ← F_{p,i*}   for all p ∈ Q

4. While |S| &lt; K:
   a. For each j ∉ S, compute marginal gains:
        Δ_c(j) = Σ_{p∈Q} (max{F_{p,j}, M_p} − M_p)
        Δ_j(j) = Σ_{i∈S} J_{i,j}
        Δ_r(j) = R_j
        Δ(j)   = λ_c·Δ_c + λ_j·Δ_j + λ_r·Δ_r
   b. j* ← argmax_j Δ(j)
   c. If Δ(j*) ≤ 0, break
   d. S ← S ∪ {j*}
      M_p ← max{M_p, F_{p,j*}}  for all p

Phase 2: Local Swap Refinement
------------------------------
Repeat up to τ passes:
  For each i_out ∈ S, j_in ∉ S:
    S' = (S \ {i_out}) ∪ {j_in}
    If score(S') &gt; score(S):
      S = S'
      restart pass
  If no improvement in pass, stop
```
<hr>
<h2>Results</h2>

Method | Spider Top-2 (P/R/F1) | Bird Top-2 (P/R/F1) | Spider Top-5 (P/R/F1) | Bird Top-5 (P/R/F1) | Spider Top-10 (P/R/F1) | Bird Top-10 (P/R/F1) | Runtime
-- | -- | -- | -- | -- | -- | -- | --
Contriever (no re-rank) | 76.2 / 72.7 / 74.0 | 65.0 / 59.4 / 61.6 | 39.8 / 93.3 / 55.5 | 37.0 / 82.9 / 50.8 | 20.6 / 96.3 / 33.8 | 21.1 / 93.9 / 34.3 | —
JAR + ILP | 82.8 / 85.8 / 87.1 | 79.1 / 73.1 / 75.3 | 39.9 / 93.7 / 55.7 | 40.3 / 89.9 / 55.2 | 20.7 / 96.6 / 33.9 | 21.5 / 95.4 / 34.9 | 1 h 06 m 55 s
JAR + Greedy (pure greedy, no local swapping) | 86.9 / 82.4 / 84.2 | 76.1 / 69.3 / 72.0 | 40.1 / 93.1 / 55.7 | 40.2 / 90.7 / 55.2 | 20.2 / 93.5 / 33.0 | 21.4 / 96.1 / 34.8 | 7 m 37 s


<p><a href="https://claude.ai/public/artifacts/29137249-ccf8-4b1b-8fd2-b15dede7948">Interactive dashboard with full comparisons</a></p>
<pre><code></code></pre></body></html>````markdown
Hello @<author>,

Thank you for sharing your work—I had a lot of fun reading your paper and appreciate the novel ideas you’ve brought to the table.

I noticed that one of the limitations you mention is the scalability bottleneck of the MIP-based re-ranking approach: as the number of candidate tables grows, the MILP solve becomes infeasible. To address this, I’ve implemented a **Greedy + Local-Swap** heuristic that achieves nearly **10× faster** runtimes while maintaining comparable end-to-end performance.

---

## Greedy + Local-Swap Heuristic

This method replaces the global MILP solve with a two-phase procedure:

1. **Greedy Selection**: incrementally build the table set by picking the candidate with the highest marginal gain.  
2. **Local Swap Refinement**: perform pairwise exchanges to recover from greedy mistakes.

Both phases optimize the same weighted objective over three signals:

$\lambda_c \sum_{p\in Q}\max_{i\in S}F_{p,i} + \lambda_j \sum_{i<j,\,i,j\in S}J_{ij} + \lambda_r \sum_{i\in S}R_i$

Where:

- **Fine-grained coverage**: $F_{p,i} \in [0,1]$ is the maximum column-similarity between sub-question $p$ and table $i$  
- **Join compatibility**: $J_{ij} = \max_{k,\ell} C^{(i,j)}_{k,\ell}$ measures how well tables $i$ and $j$ can join  
- **Coarse relevance**: $R_i$ is the retrieval score (e.g., BM25 or bi-encoder) of table $i$

---

### Pseudo-Algorithm

```text
Input:
  T = {1,…,M}        ← candidate tables  
  Q = {1,…,P}        ← decomposed sub-questions  
  K                  ← max tables to select  
  λ_c, λ_j, λ_r      ← weights  
  τ                  ← max swap passes

Phase 1: Greedy Selection
-------------------------
1. For each i ∈ T:
     G_i ← λ_c * Σ_{p∈Q} F_{p,i}  +  λ_r * R_i

2. i* ← argmax_i G_i

3. S ← {i*}
   M_p ← F_{p,i*}   for all p ∈ Q

4. While |S| < K:
   a. For each j ∉ S, compute marginal gains:
        Δ_c(j) = Σ_{p∈Q} (max{F_{p,j}, M_p} − M_p)
        Δ_j(j) = Σ_{i∈S} J_{i,j}
        Δ_r(j) = R_j
        Δ(j)   = λ_c·Δ_c + λ_j·Δ_j + λ_r·Δ_r
   b. j* ← argmax_j Δ(j)
   c. If Δ(j*) ≤ 0, break
   d. S ← S ∪ {j*}
      M_p ← max{M_p, F_{p,j*}}  for all p

Phase 2: Local Swap Refinement
------------------------------
Repeat up to τ passes:
  For each i_out ∈ S, j_in ∉ S:
    S' = (S \ {i_out}) ∪ {j_in}
    If score(S') > score(S):
      S = S'
      restart pass
  If no improvement in pass, stop
````

---

## Results (obtained on a M1-Pro Macbook)

| Method                      | Spider Top-2 (P/R/F1) | Bird Top-2 (P/R/F1) | Spider Top-5 (P/R/F1) | Bird Top-5 (P/R/F1) | Spider Top-10 (P/R/F1) | Bird Top-10 (P/R/F1) | Runtime       |
| --------------------------- | :-------------------: | :-----------------: | :-------------------: | :-----------------: | :--------------------: | :------------------: | ------------- |
| **Contriever** (no re-rank) |   76.2 / 72.7 / 74.0  |  65.0 / 59.4 / 61.6 |   39.8 / 93.3 / 55.5  |  37.0 / 82.9 / 50.8 |   20.6 / 96.3 / 33.8   |  21.1 / 93.9 / 34.3  | —             |
| **JAR + ILP**               |   82.8 / 85.8 / 87.1  |  79.1 / 73.1 / 75.3 |   39.9 / 93.7 / 55.7  |  40.3 / 89.9 / 55.2 |   20.7 / 96.6 / 33.9   |  21.5 / 95.4 / 34.9  | 1 h 06 m 55 s |
| **JAR + Greedy** (ours)     |   86.9 / 82.4 / 84.2  |  76.1 / 69.3 / 72.0 |   40.1 / 93.1 / 55.7  |  40.2 / 90.7 / 55.2 |   20.2 / 93.5 / 33.0   |  21.4 / 96.1 / 34.8  | **7 m 37 s**  |

You can find a summary in this dashboard made by Claude using the results I've obtained: https://claude.ai/public/artifacts/29137249-ccf8-4b1b-8fd2-b15dede7948


If you have any questions, suggestions or would like to talk about potential collaborations, you can reach out via my email allaa.boutaleb@lip6.fr, I've started my PhD in Paris, France @ Sorbonne University. My subject of research is table representation learning for data discovery. It's only been a few months so I would love to exchange ideas if possible :) Cheers!